src/main/java/ca/qc/ircm/proview/user/web/SignoutFilter.java
/*
 * Copyright (c) 2006 Institut de recherches cliniques de Montreal (IRCM)
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

package ca.qc.ircm.proview.user.web;

import ca.qc.ircm.proview.security.AuthenticationService;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.web.filter.GenericFilterBean;

import java.io.IOException;

import javax.inject.Inject;
import javax.servlet.FilterChain;
import javax.servlet.ServletException;
import javax.servlet.ServletRequest;
import javax.servlet.ServletResponse;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

/**
 * Allows users to sign out.
 */
public class SignoutFilter extends GenericFilterBean {
  public static final String BEAN_NAME = "SignoutFilter";
  public static final String SIGNOUT_URL = "/signout";
  private static final Logger logger = LoggerFactory.getLogger(SignoutFilter.class);
  @Inject
  private AuthenticationService authenticationService;

  public SignoutFilter() {
  }

  protected SignoutFilter(AuthenticationService authenticationService) {
    this.authenticationService = authenticationService;
  }

  @Override
  public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
      throws IOException, ServletException {
    HttpServletRequest httpRequest = (HttpServletRequest) request;
    HttpServletResponse httpResponse = (HttpServletResponse) response;
    if (httpRequest.getServletPath().equals(SIGNOUT_URL)) {
      logger.debug("sign out user {}", httpRequest.getUserPrincipal());
      authenticationService.signout();
      httpResponse.sendRedirect(httpRequest.getContextPath());
    } else {
      chain.doFilter(httpRequest, response);
    }
  }
}
